requirements.txt
google-cloud-discoveryengine
fastapi
uvicorn
pydantic
google-cloud-firestore
google-genai>=1.0.0
python-dotenv